move text to extends tool class
